+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * changed time display formatting
 * changed prompt from "<" and ">" to ";" and "*"
 * changed empty precision input to reset to default
+* check for overflow when applying offset instead of saturating
 * msrv bumped to `1.66.1`
 
 ## [0.9.2] - 2022-09-05